third_party/WebKit/LayoutTests/virtual/threaded/fast/compositorworker/request-animation-frame.html

<!DOCTYPE html>
<script src="../../../../resources/js-test.js"></script>
<script>
var jsTestIsAsync = true;
description('Tests that requestAnimationFrame is invoked with correct timestamp on CompositorWorker.');
if (window.internals)
    internals.settings.setCompositorWorkerEnabled(true);

var cw = new CompositorWorker('resources/request-animation-frame.js');
var startTime = performance.now();
var rafTime;
var endTime;
cw.onmessage = function(e) {
    rafTime = e.data;
    endTime = performance.now();
    shouldBeGreaterThanOrEqual('rafTime', 'startTime');
    shouldBeGreaterThanOrEqual('endTime', 'rafTime');
    testPassed('Received reply from CompositorWorker.');
    finishJSTest();
}
</script>
