third_party/WebKit/LayoutTests/virtual/threaded/fast/compositorworker/request-animation-frame-expected.txt

Tests that requestAnimationFrame is invoked with correct timestamp on CompositorWorker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rafTime is >= startTime
PASS endTime is >= rafTime
PASS Received reply from CompositorWorker.
PASS successfullyParsed is true

TEST COMPLETE
